fix a small memory leak
When it encountered an eof, sc_get_next_frame would set the state to
STATE_CLOSED. But that makes sc_close think that the stream is already
closed so it doesn't do any cleanup. Introduce a new STATE_EOF for the
eof case, allowing sc_close to actually work properly.